Address issue #18 by checking if the file exists before unlinking it

+++ src/Brainy/Templates/Template.php
@@ -327,7 +327,9 @@ PHPDOC;
         try {
             if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN') {
                 // remove original file
-                unlink($_filepath);
+                if (file_exists($_filepath)) {
+                    unlink($_filepath);
+                }
                 // rename tmp file
                 $success = rename($_tmp_file, $_filepath);
             } else {
